Move Admin from components to pages
